Fix asciidoctor-diagram jekyll image output issue
Configure the imagesoutdir attribute on the Asciidoctor convert call.

This allow Asciidoctor Diagram to output to the correct folder, but jekyll
has the following lifecycle: read, generate, render, cleanup, write.

The diagram images are written in the render phase, then jekyll cleans the _site
folder and the 'known' jekyll data is written. This end up cleaning all the images
created by Asciidoctor-Diagrams.

Monkey patch the Jekyll Site cleaner to not delete the _site dir. In dev this
needs to be done manually.